packages/llm/llamaspeak/requirements.txt
nvidia-riva-client
websockets